requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
gunicorn
django-cors-headers
psycopg2-binary



